Refactor ShoppingList using custom hooks
